updated code and trying to build

+++ src/i2c_temp.h
@@ -35,7 +35,7 @@ void i2ctemp_On();
 void i2ctemp_Sleep();
 void i2c_read_tempreg(I2C_TypeDef *i2c, uint8_t slaveAddr, uint8_t reg_addr);
 void i2c_write_tempreg(I2C_TypeDef *i2c, uint8_t slaveAddr, uint8_t reg_addr);
-int get_temp_value();
+float get_temp_value();
 
 #endif
